Don't send as many gratuitous arps
The DHCP client sends one when it gets a lease and then the stack sends
them periodically thereafter based on the value of
arpGRATUITOUS_ARP_PERIOD.

We seem to send far more than my very small sampling of other stacks so
this patch opens the period out to a reasonably large value that still fits
in the smallest TickType_t we're likely to encounter on a small platform.

Opening this value out means that the CPU can spend more time in sleep
modes.

900,000ms is 15 minutes and this fits in a uint16_t TickType_t on an 8-bit
avr8 with ~15ms ticks.

Signed-off-by: Andy Bennett <andyjpb@ashurst.eu.org>