Increase the ipTCP_TIMER_PERIOD_MS to allow sleep modes when there's unlikely to be anything to do
On a fairly quiescent TCP stack with no open TCP connections, the low value
for ipTCP_TIMER_PERIOD_MS prevents the CPU from staying in sleep modes
despite the task finding nothing to do.

ipTCP_TIMER_PERIOD_MS is the longest time that will be allowed between
checks: if the stack knows there is definitely work to do before then, it
will set the timer for a smaller period.

Here we extend the timeout that is used by FreeRTOS_Sockets.c.
FreeRTOS_IP.c still uses a value of 1000ms and that value is used when the
TCP stack is first initialised.

Signed-off-by: Andy Bennett <andyjpb@ashurst.eu.org>